+++ fluid/WindowType.cxx
@@ -94,7 +94,7 @@ void set_preferences_window(){
 
 void show_preferences_cb(Widget *, void * tabnum) {
   set_preferences_window();
-  int n = (int) (long) tabnum;
+  int n = (int) (uintptr_t) tabnum;
   if (n>=0 && n<3)
       pref_tabs->value(n);
   preferences_window->show();
